Handle out-of-date raw metadata
This changes `raw_metadata()` to accept a key path and return the data
it finds.

It also changes the function to try `ffprobe` again if the metadata
doesn’t contain the key path. This seems like a pretty small edge case
that will only come up when moving from one version of yanki to another.