Add `--reprocess` and configure `cache_path` on Deck
Previously, `cache_path` was passed to the parser and was passed along,
unused, to the `Video`. This removes it from the parser and spec objects
and inserts it when the `Deck` is created.

This adds a `--reprocess` option to force reproccessing videos (without
re-downloading them). This is mostly useful if I change how filters are
implemented, since their functioning isn’t represented in the cached
file name.