Simplify ffmpeg calls to only run once
Previously, we used the `-copyts` option when we wanted to clip off the
start and end of a video. This switches to calculating the duration of
the video, so we no longer have to use `-copyts`. This means that we no
longer need to have a second pass with ffmpeg.