Update default note_id format
My recent changes to timestamp normalization have changed a bunch of
`note_id`s, so it’s a good time to change the default.

This changes the time format to consistently be `'%0.06f'`, since ffmpeg
supports a `us` suffix for times.

This removes the "__" after the `{deck_id}`; that was there because we
used to include the deck_id as a separate parameter to the GUID
generation function.

This removes `{direction}` since we can now change the direction of a
card after it has been created.

This also changes the note_id format for all ASL decks, since we’re
going through the pain of updating them anyway.