Expand list of invalid characters in file names
There are issues with certain characters in file names (see issue #8 —
Check how Anki handles special characters in `[sound: ...]`).

I’ve listed my best guess at which characters are a problem in
`FILENAME_ILLEGAL_CHARS`, added escaping to prevent then, and added a
check to ensure that none made it into the video ID, which is embedded
directly in a file name.