`yanki.utils` and `atomic_open()`
This moves some utility functions out of `yanki.video`, and creates a
new `atomic_open()` function that generates a temporary file, then
moves it into place atomically.

This ensures that we don’t end up with a corrupt file, e.g. of metadata,
if it is interupted in the middle of writing the file.

Fixes: #33 — Create cache files atomically