Move config directive handling into `NoteConfig`
This switches to expecting a consistent syntax for config directives and
raising an error if it finds something that matches the syntax but is
not a known config directive.

It also moves the knowledge of what the config directives are (other
than `title`) into `NoteConfig`.

This makes `NoteConfig` a `dataclass`, and generates a frozen version of
it to use with `NoteSpec`.