ASL lesson 7 deck: update comments
Note that CUP was in lesson 7.

Remove comment explaining how to generate groups from the lesson HTML.
Here is the current best version:

```javascript
Array.from(window.getSelection().getRangeAt(0).cloneContents().querySelectorAll('a[href]')).map(a => {
    var result = /\/([^\/]+)\.html?$/i.exec(a.href);
    var word = "source";
    if (result) {
        word = result[1].toUpperCase();
    }
    return "\ngroup:\n  more: +rst:`" + word + " <" + a.href + ">`_\n";
}).join("")
```

To use, select the vocabulary list on the lesson page, then open the
console and paste the above code. It will detect all the links and turn
them into groups printed to the console.