ASL decks: add lesson 19
This moves a few signs from other decks.

Here is the latest update to the script that generates groups from the
lesson vocabulary HTML:

```javascript
Array.from(window.getSelection().getRangeAt(0).cloneContents().querySelectorAll('a[href]')).map(a => {
    var result = /\/([^\/]+)\.html?$/i.exec(a.href);
    var word = "source";
    if (result) {
        word = result[1].toUpperCase();
    }
    return "\ngroup:\n  more: +md:[" + word + "](" + a.href + ")\n";
}).join("")
```

To use, select the vocabulary list on the lesson page, then open the
console and paste the above code. It will detect all the links and turn
them into groups printed to the console.