ASL decks: make gloss more consistent
The big thing here is that the all-capital gloss is the essentially the
name of the sign. Lifeprint is note entirely consistent about this, so
in a few places I left multiple all-caps versions.

I also started switching many non-gloss translations to not use hyphens.
For example the sign NYB! can be translated to English as “none of your
business” — that’s not gloss.

This also tweaks a few things to better match the Lifeprint web site and
trims a few videos.