Fix: collect more_info from clipped videos
Previously the cache key for more_info was just the video ID. This adds
the video clip to the cache key so that crops that are different in
different parts of the video are properly supported.

Technically the above is correct, but really the more important
consequence of this change is that it avoids multiple `async` functions
working on the same more_info concurrently. When I add the `scdet`
filter concurrent processes working on big videos really bogs things
down (e.g. for the ASL numbers decks).