`Video`: Don’t check metadata `mtime`
Previously we refreshed the metadata (`ffprobe` results) if it was older
than the raw video. This would interfere with touching files in the
cache to track what is actually being used, so now we no longer check
the `mtime`.

We don’t have a re-download option, but if we did it would need to
remove metadata as well as a few other files.