`Cache`: add `async_lock_limit` to limit open files
A slow cached `async` function that’s run on a bunch of items (e.g.
videos) can cause a “too many open files error” pretty easily, as each
instance of the function holds open a the lock file on the cache.

This adds a simple way to limit the number of cached `async` functions
run at once.

This is likely useful for avoiding other constraints as well.