Web UI: fix viewer to not extend off edge of page
The `min-width` was set to the width of the window even though the
viewer sometimes has the directory sidebar next to it. Without the
`min-width` property the viewer’s minimum width is the width of its
contents, so it’s required to be present to shrink the title below the
length of its text.

There was also a `min-width` property on the cards, but it does not seem
to be necessary.